% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrate_half_degree.r
\name{agrate_half_degree}
\alias{agrate_half_degree}
\title{agrate_half_degree}
\usage{
agrate_half_degree(input, ag_by = "region")
}
\arguments{
\item{input}{either a raster of half degree cells or a data frame with 259200 cells in the order output by raster}

\item{ag_by}{'country' or 'region' later can offer other options, e.g. 'continent'}
}
\description{
to agregate a half degree grid to country or region, gives n and mean
Probably should go in another package, but here for a start.
}
\examples{
#agrate_half_degree()
ec_out_potato <- raster(system.file("extdata", "ec_out_potato.grd", package = "climcropr"))
agrate_half_degree(ec_out_potato)

}
