% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rast_change.r
\name{calc_rast_change}
\alias{calc_rast_change}
\title{calc_rast_change
calculating maps of change, future-present is what we want, gain=1, loss=-1.
this function not really necessary can just do simple subtraction}
\usage{
calc_rast_change(now, fut)
}
\arguments{
\item{now}{raster of now}

\item{fut}{raster of future}
}
\value{
raster where +1 = gain, -1 = loss
}
\description{
present		future 	f-p
0		0	0
0		1	1
1		0	-1
1		1	0
}
