% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_plots.r
\name{cr_plot_cycle}
\alias{cr_plot_cycle}
\title{cr_plot_cycle plot length of crop cycle from ecocrop database}
\usage{
cr_plot_cycle(cropnames = c("maize", "rice"))
}
\arguments{
\item{cropnames}{name or vector of crop names}
}
\value{
ggplot2 object
}
\description{
for a single or vector of crops
}
\examples{
cropnames <- c('broom-corn','maize','rice','potato','soya bean','sugarcane')
cr_plot_cycle(cropnames)
cr_plot_cycle(c('asparagus','maize','celery','rape','cabbage'))
}
