% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_suit_simpler.R
\name{cr_suit_simpler}
\alias{cr_suit_simpler}
\title{cr_suit_simpler a simpler than ecocrop suitability function along similar lines}
\usage{
cr_suit_simpler(crop, tmin, tavg, use_tmax = FALSE, tmax = NULL, prec,
  rainfed)
}
\arguments{
\item{crop}{an ecocrop cropname or object}

\item{tmin}{monthly tmin for a site}

\item{tavg}{monthly tavg for a site}

\item{use_tmax}{option to use tmax rather than tavg}

\item{tmax}{monthly tmax for a site}

\item{prec}{monthly precipitation for a site}

\item{rainfed}{FALSE for irrigated}
}
\value{
1/0 for suitable/unsuitable
}
\description{
uses crop@GMIN / 30 as minimum months of crop cycle needed
temp_suit : min cycle consecutive months where tavg is within crop@TMIN, crop@TMAX
kill_temp : tmin-5 > crop@KTMP
precip : total precip within growing cycle should be within crop@RMIN, crop@RMAX
}
\examples{
crop <- getCrop('potato')
cr_suit_simpler(crop, 5:16, 15:26, runif(12)*100, rainfed=TRUE)
}
