% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_suit_simpler.R
\name{cr_suit_simpler_month}
\alias{cr_suit_simpler_month}
\title{cr_suit_simpler_month}
\usage{
cr_suit_simpler_month(crop, tmin, tavg, use_tmax, tmax, prec, rainfed)
}
\arguments{
\item{crop}{an ecocrop cropname or object}

\item{tmin}{single tmin value for a site}

\item{tavg}{single tavg value for a site}

\item{use_tmax}{option to use tmax rather than tavg}

\item{tmax}{monthly tmax for a site}

\item{prec}{single precipitation value for a site}

\item{rainfed}{FALSE for irrigated}
}
\value{
dataframe of suitabilities by month and attribute
}
\description{
a simpler version of the ecocrop model that assesses temperature against the minimum growth cycle
and rainfall against the maximum growth cycle.
returns suitabilities by month and attribute
}
\examples{
crop <- getCrop('potato')
cr_suit_simpler_month(crop, 5:16, 15:26, runif(12)*100, rainfed=TRUE)

}
