% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_suit_simpler.R
\name{cr_suit_simpler_month_test}
\alias{cr_suit_simpler_month_test}
\title{cr_suit_simpler_month_test}
\usage{
cr_suit_simpler_month_test(crop, tmin, tavg, prec, rainfed)
}
\arguments{
\item{crop}{an ecocrop cropname or object}

\item{tmin}{single tmin value for a site}

\item{tavg}{single tavg value for a site}

\item{prec}{single precipitation value for a site}

\item{rainfed}{FALSE for irrigated}
}
\value{
dataframe of suitabilities by month and attribute
}
\description{
BEWARE THIS A TEST VERSION THAT USED BOTH GMIN & GMAX, THE ALL IS NOT USEFUL
returns suitabilities by month and attribute
}
\examples{
crop <- getCrop('potato')
cr_suit_simpler_month_test(crop, 5:16, 15:26, runif(12)*100, rainfed=TRUE)

}
