% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_crop}
\alias{df_crop}
\title{df_crop}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 2 columns.}
\usage{
df_crop
}
\description{
\describe{
  \item{code}{MIRCA2000 Crop Code}
  \item{name}{Crop Name in English, \emph{e.g.}, "Potatoes"}
}
}
\keyword{datasets}
