% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_mirca}
\alias{df_mirca}
\title{df_mirca}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 3 columns.}
\usage{
df_mirca
}
\description{
\describe{
  \item{mirca_code}{MIRCA2000 Crop Code}
  \item{mirca_name}{Mirca Name in English, \emph{e.g.}, "Potatoes"}
  \item{ecocrop_name}{ecocrop Name in English, \emph{e.g.}, "Potato"}
}
}
\keyword{datasets}
