% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecocrop.r
\name{ecocrop}
\alias{ecocrop}
\title{ecocrop}
\usage{
ecocrop(crop, tmin, tavg, prec, rainfed)
}
\arguments{
\item{crop}{an ecocrop cropname or object}

\item{tmin}{single tmin value for a site}

\item{tavg}{single tavg value for a site}

\item{prec}{single precipitation value for a site}

\item{rainfed}{FALSE for irrigated}
}
\description{
dismo::ecocrop function, copied here temporarily to see if we can extract parts
and/or increase flexibility.
With added comments.
}
\details{
adds suitability for a single pixel to the ecocrop object

This is from .doEcocrop, some additional data setup occurs in dismo::ecocrop and .ecoSpat()

source : https://r-forge.r-project.org/scm/viewvc.php/pkg/dismo/R/ecocrop.R?view=markup&root=dismo
}
\examples{
crop <- getCrop('potato')
ecocrop(crop, 5:16, 15:26, runif(12)*100, rainfed=TRUE)
crop <- getCrop('Hot pepper')
ecocrop(crop, 5:16, 15:26, rainfed=FALSE)
}
