% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirca2ecocrop.r
\name{ecocrop2mirca}
\alias{ecocrop2mirca}
\title{ecocrop2mirca converts from mirca name to the closest ecocrop name}
\usage{
ecocrop2mirca(name)
}
\arguments{
\item{name}{an ecocrop cropname}
}
\value{
a translated cropname
}
\description{
ecocrop2mirca converts from mirca name to the closest ecocrop name
}
\examples{
ecocrop2mirca('potato')
ecocrop2mirca('broom-corn')
}
