% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecocrop.r
\name{get_ecocrop}
\alias{get_ecocrop}
\title{get_ecocrop}
\usage{
get_ecocrop(cropname)
}
\arguments{
\item{cropname}{an ecocrop cropname}
}
\description{
get new ecocrop entry for a crop
}
\details{
data scraped from FAO website 2017
}
\examples{
potato <- get_ecocrop('potato')
}
