% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mirca.r
\name{get_mirca}
\alias{get_mirca}
\title{Get a MIRCA2000 layer}
\usage{
get_mirca(cropname, rainfed = TRUE, plot = TRUE, cache = TRUE)
}
\arguments{
\item{cropname}{a MIRCA2000 crop name, see \code{\link{df_crop}} for a full
list}

\item{rainfed}{default=TRUE, FALSE for irrigated}

\item{plot}{default=TRUE, plot the resulting raster}

\item{cache}{default=FALSE, TRUE to save MIRCA2000 files locally for future
use with \code{climcropr}}
}
\value{
A \code{\link[raster]{raster}} object of MIRCA2000 crop data and
optionally a gzipped file of the original data saved to local disk for future
use with \code{climcropr}.
}
\description{
Download MIRCA2000 global data of monthly irrigated and rainfed
crop areas around the year 2000 (MIRCA2000)
}
\note{
The data used in this package are downloaded from:

\url{ftp://ftp.rz.uni-frankfurt.de/pub/uni-frankfurt/physische_geographie/hydrologie/public/data/MIRCA2000/harvested_area_grids/}
}
\examples{
rst_mir <- get_mirca("potatoes", rainfed = TRUE)

}
\references{
The documentation for irrigated crops was published as Frankfurt Hydrology
Paper 06:

Portmann, F., Siebert, S., Bauer, C. & Döll, P. (2008): Global data set of
monthly growing areas of 26 irrigated crops. Frankfurt Hydrology Paper 06,
Institute of Physical Geography, University of Frankfurt, Frankfurt am Main,
Germany.

The documentation for both rainfed and irrigated crops is published in a
peer-reviewed journal:

Portmann, F. T., Siebert, S. & Döll, P. (2010): MIRCA2000 – Global monthly
irrigated and rainfed crop areas around the year 2000: A new high-resolution
data set for agricultural and hydrological modeling, Global Biogeochemical
Cycles, 24, GB 1011, doi:10.1029/2008GB003435.

For Frankfurt Hydrology Paper 09, the documentation for irrigated crops was
updated especially with respect to data quality and scaling. Data quality
parameters by country and unit, and crop calendars for rainfed and irrigated
crops for each unit are now available, too:

Portmann, F.T. (2011): Global estimation of monthly irrigated and rainfed
crop areas on a 5 arc-minute grid. Frankfurt Hydrology Paper 09, Institute of
Physical Geography, University of Frankfurt, Frankfurt am Main, Germany.
}
