% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirca2ecocrop.r
\name{mirca2ecocrop}
\alias{mirca2ecocrop}
\title{mirca2ecocrop converts from mirca name to the closest ecocrop name}
\usage{
mirca2ecocrop(name)
}
\arguments{
\item{name}{a mirca cropname}
}
\value{
a translated cropname
}
\description{
mirca2ecocrop converts from mirca name to the closest ecocrop name
}
\examples{
mirca2ecocrop('potatoes')
mirca2ecocrop('sorghum')
}
