% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_crop_lims.r
\name{set_crop_lims}
\alias{set_crop_lims}
\title{set_crop_lims}
\usage{
set_crop_lims(GMIN, GMAX, KTMP = 0, TMIN, TMAX, RMIN, RMAX)
}
\arguments{
\item{GMIN}{minimum growth cycle in days}

\item{GMAX}{maximum growth cycle in days}

\item{KTMP}{killing temperature degrees C}

\item{TMIN}{min temperature degrees C}

\item{TMAX}{max temperature degrees C}

\item{RMIN}{min rainfall needed in growth cycle mm}

\item{RMAX}{max rainfall tolerated in growth cycle mm}
}
\description{
create a crop object with specified limits to allow suitability modelling
}
\examples{
made_up_crop <- set_crop_lims(GMIN=90, GMAX=120, KTMP=0, TMIN=10, TMAX=35, RMIN=500, RMAX=2000)
#test running suitability for small area
#subset climate data to make faster run
st_clim2 <- raster::crop(st_clim, extent(-10,10,45,65))
rastsuit <- ecocrop_a_raster(made_up_crop, st_clim2, simpler=TRUE)
}
