% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suit_soil_ph.r
\name{suit_soil_ph}
\alias{suit_soil_ph}
\title{suit_soil_ph}
\usage{
suit_soil_ph(cropname, phmin = NULL, phmax = NULL, phopmn = NULL,
  phopmx = NULL, suit_or_opt = "suit1_opt2", filename = "", plot = TRUE)
}
\arguments{
\item{cropname}{name of the crop}

\item{suit_or_opt}{'suit', 'opt', 'suit1_opt2'}

\item{filename}{optional output file}

\item{plot}{whether to plot result}
}
\description{
generate a global suitability (0/1) raster for a crop based on it's ph limits from ecocrop
and harmonised world soil database data at 0.05 degree resolution
}
\details{
TODO offer option to pass ph limits as args rather than getting from ecocrop
TODO offer option to use optimal rather than absolute ph limits from ecocrop
}
\examples{
#takes a few mins to run
#rast_potato_suit_ph <- suit_soil_ph('potato')


}
