% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_pos.r
\name{true_neg}
\alias{true_neg}
\title{true_neg simple method to calculate true negative rate between an observed and predicted raster}
\usage{
true_neg(obs, pred)
}
\arguments{
\item{obs}{raster of observed values}

\item{pred}{raster of predicted values}
}
\value{
true negative rate between 0 & 1
}
\description{
treats cells that are 0 in both layers as true negative
}
\examples{
crop <- 'maize'
obs <- raster::subset(st,paste0(crop,'_mirca'))
pred <- raster::subset(st,paste0(crop,'_suitsimp'))
tneg <- true_neg(obs=obs, pred=pred)
}
