% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_t_thresh.r
\name{within_t_thresh}
\alias{within_t_thresh}
\title{within_t_thresh}
\usage{
within_t_thresh(crop, month_tmin = NULL, month_tmax = NULL,
  min_or_max = "both")
}
\arguments{
\item{crop}{the crop either a name or ecocrop object}

\item{month_tmin}{min temp by month}

\item{month_tmax}{max temp by month}

\item{min_or_max}{'both', min','max'}
}
\description{
are a crop's temperature thresholds satisfied by the passed monthly temperature data ?
TODO currently doesn't take into account growing season length but does need to.
}
\examples{
#0
within_t_thresh('potato', month_tmin = rep(0,12), month_tmax = rep(50,12))
#1
within_t_thresh('potato', month_tmin = rep(20,12), month_tmax = rep(20,12))
#just min or max. e.g. specify max and it ignores min
within_t_thresh('potato', month_tmin = rep(0,12), month_tmax = rep(20,12), min_or_max='max')
}
