% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mirca.r
\name{get_mirca}
\alias{get_mirca}
\title{get_mirca}
\usage{
get_mirca(cropname, rainfed = TRUE, plot = TRUE)
}
\arguments{
\item{cropname}{a Mirca cropname}

\item{rainfed}{default=TRUE, FALSE for irrigated}

\item{plot}{whether to plot the raster}
}
\description{
get a mirca layer
}
\details{
data from
ftp://ftp.rz.uni-frankfurt.de/pub/uni-frankfurt/physische_geographie/hydrologie/public/data/MIRCA2000/harvested_area_grids/
}
\examples{
rst_mir <- get_mirca(name_mir, rainfed = TRUE)
raster::plot(rst_mir)
}
