% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suit_soil_ph.r
\name{suit_soil_ph}
\alias{suit_soil_ph}
\title{suit_soil_ph}
\usage{
suit_soil_ph(cropname, phmin = NULL, phmax = NULL, phopmn = NULL,
  phopmx = NULL, suit_or_opt = "suit1_opt2", filename = "", plot = TRUE)
}
\arguments{
\item{cropname}{name of the crop}

\item{phmin}{min ph from ecocrop}

\item{phmax}{max ph from ecocrop}

\item{phopmn}{min optimal ph from ecocrop}

\item{phopmx}{max optimal ph from ecocrop}

\item{suit_or_opt}{'suit', 'opt', 'suit1_opt2'}

\item{filename}{optional output file}

\item{plot}{whether to plot result}
}
\description{
generate a global suitability (0/1) raster for a crop based on it's ph limits from ecocrop
and regridded harmonised world soil database v1.2 at 0.05 degree resolution
}
\details{
Wieder, W.R., J. Boehnert, G.B. Bonan, and M. Langseth. 2014. Regridded Harmonized World Soil Database v1.2.
Data set. Available on-line [http://daac.ornl.gov] from Oak Ridge National Laboratory Distributed
Active Archive Center, Oak Ridge, Tennessee, USA. http://dx.doi.org/10.3334/ORNLDAAC/1247.
}
\examples{
#takes a few mins to run
#rast_potato_suit_ph <- suit_soil_ph('potato')


}
