% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_pos.r
\name{true_pos}
\alias{true_pos}
\title{true_pos simple method to calculate true positive rate between an observed and predicted raster where each is greater than zero}
\usage{
true_pos(obs, pred)
}
\arguments{
\item{obs}{raster of observed values}

\item{pred}{raster of predicted values}
}
\value{
true positive rate between 0 & 1
}
\description{
treats cells that are greater than 0 in both layers as true positive
}
\examples{
crop <- 'maize'
obs <- raster::subset(st,paste0(crop,'_mirca'))
pred <- raster::subset(st,paste0(crop,'_suitsimp'))
tpos_s <- true_pos(obs=obs, pred=pred)
}
